backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-jose
bcrypt
python-multipart
python-dotenv
